Added hx5_d20_rev2 left URDF and updated moveit_config
