+++ robotis_hand_bringup/launch/hx5_d20_left_gazebo.launch.py
@@ -28,7 +28,11 @@ from launch.actions import SetEnvironmentVariable
 from launch.event_handlers import OnProcessExit
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 from launch.substitutions import LaunchConfiguration
+from launch.substitutions import PathJoinSubstitution
+from launch.substitutions import Command
 from launch_ros.actions import Node
+from launch_ros.descriptions import ParameterValue
+
 import xacro
 
 
@@ -72,24 +76,33 @@ def generate_launch_description():
 
     model = LaunchConfiguration('model')
 
-    xacro_file = os.path.join(
+    xacro_file = PathJoinSubstitution([
         robotis_hand_description_path,
         'urdf',
         model,
-        'hx5_d20_left.xacro',
+        'hx5_d20_left.xacro'
+    ])
+
+    robot_desc_command = Command([
+        'xacro',
+        ' ',
+        xacro_file,
+        ' ',
+        'use_sim:=true'
+    ])
+
+    robot_desc_content = ParameterValue(
+        value=robot_desc_command,
+        value_type=str
     )
-
-    doc = xacro.process_file(xacro_file, mappings={'use_sim': 'true'})
-
-    robot_desc = doc.toprettyxml(indent='  ')
-
-    params = {'robot_description': robot_desc}
 
     node_robot_state_publisher = Node(
         package='robot_state_publisher',
         executable='robot_state_publisher',
         output='screen',
-        parameters=[params],
+        parameters=[
+            {'robot_description': robot_desc_content}
+        ]
     )
 
     gz_spawn_entity = Node(
@@ -98,7 +111,7 @@ def generate_launch_description():
         output='screen',
         arguments=[
             '-string',
-            robot_desc,
+            robot_desc_command,
             '-x',
             '0.0',
             '-y',
