+++ robotis_hand_moveit_config/config/hx5_d20_left/hx5_d20_left.srdf
@@ -1,0 +1,107 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--This does not replace URDF, and is not an extension of URDF.
+    This is a format for representing semantic information about the robot structure.
+    A URDF file must exist for this robot as well, where the joints and the links that are referenced are defined
+-->
+<robot name="hx5_d20_left">
+    <!--GROUPS: Representation of a set of joints and links. This can be useful for specifying DOF to plan for, defining arms, end effectors, etc-->
+    <!--LINKS: When a link is specified, the parent joint of that link (if it exists) is automatically included-->
+    <!--JOINTS: When a joint is specified, the child link of that joint (which will always exist) is automatically included-->
+    <!--CHAINS: When a chain is specified, all the links along the chain (including endpoints) are included in the group. Additionally, all the joints that are parents to included links are also included. This means that joints along the chain and the parent joint of the base link are included in the group-->
+    <!--SUBGROUPS: Groups can also be formed by referencing to already defined group names-->
+    <group name="thumb">
+        <joint name="finger_l_joint1"/>
+        <joint name="finger_l_joint2"/>
+        <joint name="finger_l_joint3"/>
+        <joint name="finger_l_joint4"/>
+        <joint name="finger_end_l_joint1"/>
+    </group>
+    <group name="index">
+        <joint name="finger_l_joint5"/>
+        <joint name="finger_l_joint6"/>
+        <joint name="finger_l_joint7"/>
+        <joint name="finger_l_joint8"/>
+        <joint name="finger_end_l_joint2"/>
+    </group>
+    <group name="middle">
+        <joint name="finger_l_joint9"/>
+        <joint name="finger_l_joint10"/>
+        <joint name="finger_l_joint11"/>
+        <joint name="finger_l_joint12"/>
+        <joint name="finger_end_l_joint3"/>
+    </group>
+    <group name="ring">
+        <joint name="finger_l_joint13"/>
+        <joint name="finger_l_joint14"/>
+        <joint name="finger_l_joint15"/>
+        <joint name="finger_l_joint16"/>
+        <joint name="finger_end_l_joint4"/>
+    </group>
+    <group name="pinky">
+        <joint name="finger_l_joint17"/>
+        <joint name="finger_l_joint18"/>
+        <joint name="finger_l_joint19"/>
+        <joint name="finger_l_joint20"/>
+        <joint name="finger_end_l_joint5"/>
+    </group>
+    <!--GROUP STATES: Purpose: Define a named state for a particular group, in terms of joint values. This is useful to define states like 'folded arms'-->
+    <group_state name="init" group="thumb">
+        <joint name="finger_l_joint1" value="0"/>
+        <joint name="finger_l_joint2" value="0"/>
+        <joint name="finger_l_joint3" value="0"/>
+        <joint name="finger_l_joint4" value="0"/>
+    </group_state>
+    <group_state name="init" group="index">
+        <joint name="finger_l_joint5" value="0"/>
+        <joint name="finger_l_joint6" value="0"/>
+        <joint name="finger_l_joint7" value="0"/>
+        <joint name="finger_l_joint8" value="0"/>
+    </group_state>
+    <group_state name="init" group="middle">
+        <joint name="finger_l_joint10" value="0"/>
+        <joint name="finger_l_joint11" value="0"/>
+        <joint name="finger_l_joint12" value="0"/>
+        <joint name="finger_l_joint9" value="0"/>
+    </group_state>
+    <group_state name="init" group="ring">
+        <joint name="finger_l_joint13" value="0"/>
+        <joint name="finger_l_joint14" value="0"/>
+        <joint name="finger_l_joint15" value="0"/>
+        <joint name="finger_l_joint16" value="0"/>
+    </group_state>
+    <group_state name="init" group="pinky">
+        <joint name="finger_l_joint17" value="0"/>
+        <joint name="finger_l_joint18" value="0"/>
+        <joint name="finger_l_joint19" value="0"/>
+        <joint name="finger_l_joint20" value="0"/>
+    </group_state>
+    <!--END EFFECTOR: Purpose: Represent information about an end effector.-->
+    <end_effector name="end_effector_thumb" parent_link="finger_end_l_link1" group="thumb"/>
+    <end_effector name="end_effector_index" parent_link="finger_end_l_link2" group="index"/>
+    <end_effector name="end_effector_middle" parent_link="finger_end_l_link3" group="middle"/>
+    <end_effector name="end_effector_ring" parent_link="finger_end_l_link4" group="ring"/>
+    <end_effector name="end_effector_pinky" parent_link="finger_end_l_link5" group="pinky"/>
+    <!--DISABLE COLLISIONS: By default it is assumed that any link of the robot could potentially come into collision with any other link in the robot. This tag disables collision checking between a specified pair of links. -->
+    <disable_collisions link1="finger_l_link1" link2="finger_l_link17" reason="Never"/>
+    <disable_collisions link1="finger_l_link1" link2="finger_l_link2" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link1" link2="hx5_d20_left_base" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link10" link2="finger_l_link11" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link10" link2="finger_l_link9" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link11" link2="finger_l_link12" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link13" link2="finger_l_link14" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link13" link2="hx5_d20_left_base" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link14" link2="finger_l_link15" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link15" link2="finger_l_link16" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link17" link2="finger_l_link18" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link17" link2="finger_l_link5" reason="Never"/>
+    <disable_collisions link1="finger_l_link17" link2="hx5_d20_left_base" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link18" link2="finger_l_link19" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link19" link2="finger_l_link20" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link2" link2="finger_l_link3" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link3" link2="finger_l_link4" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link5" link2="finger_l_link6" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link5" link2="hx5_d20_left_base" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link6" link2="finger_l_link7" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link7" link2="finger_l_link8" reason="Adjacent"/>
+    <disable_collisions link1="finger_l_link9" link2="hx5_d20_left_base" reason="Adjacent"/>
+</robot>
