+++ robotis_hand_moveit_config/config/hx5_d20_left/joint_limits.yaml
@@ -1,0 +1,175 @@
+# joint_limits.yaml allows the dynamics properties specified in the URDF to be overwritten or augmented as needed
+
+# For beginners, we downscale velocity and acceleration limits.
+# You can always specify higher scaling factors (<= 1.0) in your motion requests.  # Increase the values below to 1.0 to always move at maximum speed.
+default_velocity_scaling_factor: 0.1
+default_acceleration_scaling_factor: 0.1
+
+# Specific joint properties can be changed with the keys [max_position, min_position, max_velocity, max_acceleration]
+# Joint limits can be turned off with [has_velocity_limits, has_acceleration_limits]
+joint_limits:
+  finger_l_joint1:
+    has_position_limits: true
+    min_position: -1.57
+    max_position: 1.57
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint2:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 3.14
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint3:
+    has_position_limits: true
+    min_position: -1.57
+    max_position: 0.0
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint4:
+    has_position_limits: true
+    min_position: -1.57
+    max_position: 0.0
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint5:
+    has_position_limits: true
+    min_position: -0.6
+    max_position: 0.6
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint6:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 2.0
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint7:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 1.57
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint8:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 1.57
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint9:
+    has_position_limits: true
+    min_position: -0.6
+    max_position: 0.6
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint10:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 2.0
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint11:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 1.57
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint12:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 1.57
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint13:
+    has_position_limits: true
+    min_position: -0.6
+    max_position: 0.6
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint14:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 2.0
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint15:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 1.57
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint16:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 1.57
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint17:
+    has_position_limits: true
+    min_position: -0.6
+    max_position: 0.6
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint18:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 2.0
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint19:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 1.57
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  finger_l_joint20:
+    has_position_limits: true
+    min_position: 0.0
+    max_position: 1.57
+    has_velocity_limits: true
+    max_velocity: 5.0
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+
+# Goal Tolerances for the entire planning group
+thumb:
+  goal_position_tolerance: 0.01
+  goal_orientation_tolerance: 0.01
