+++ robotis_hand_moveit_config/config/hx5_d20_left/moveit_controllers.yaml
@@ -1,0 +1,47 @@
+# MoveIt uses this configuration for controller management
+
+moveit_controller_manager: moveit_simple_controller_manager/MoveItSimpleControllerManager
+
+moveit_simple_controller_manager:
+  controller_names:
+    - thumb_controller
+    - index_controller
+    - middle_controller
+    - ring_controller
+    - pinky_controller
+
+  thumb_controller:
+    type: FollowJointTrajectory
+    joints:
+      - finger_l_joint1
+      - finger_l_joint2
+      - finger_l_joint3
+      - finger_l_joint4
+  index_controller:
+    type: FollowJointTrajectory
+    joints:
+      - finger_l_joint5
+      - finger_l_joint6
+      - finger_l_joint7
+      - finger_l_joint8
+  middle_controller:
+    type: FollowJointTrajectory
+    joints:
+      - finger_l_joint9
+      - finger_l_joint10
+      - finger_l_joint11
+      - finger_l_joint12
+  ring_controller:
+    type: FollowJointTrajectory
+    joints:
+      - finger_l_joint13
+      - finger_l_joint14
+      - finger_l_joint15
+      - finger_l_joint16
+  pinky_controller:
+    type: FollowJointTrajectory
+    joints:
+      - finger_l_joint17
+      - finger_l_joint18
+      - finger_l_joint19
+      - finger_l_joint20
