--- robotis_hand_moveit_config/config/moveit_controllers.yaml
@@ -1,47 +1,0 @@
-# MoveIt uses this configuration for controller management
-
-moveit_controller_manager: moveit_simple_controller_manager/MoveItSimpleControllerManager
-
-moveit_simple_controller_manager:
-  controller_names:
-    - thumb_controller
-    - index_controller
-    - middle_controller
-    - ring_controller
-    - pinky_controller
-
-  thumb_controller:
-    type: FollowJointTrajectory
-    joints:
-      - finger_r_joint1
-      - finger_r_joint2
-      - finger_r_joint3
-      - finger_r_joint4
-  index_controller:
-    type: FollowJointTrajectory
-    joints:
-      - finger_r_joint5
-      - finger_r_joint6
-      - finger_r_joint7
-      - finger_r_joint8
-  middle_controller:
-    type: FollowJointTrajectory
-    joints:
-      - finger_r_joint9
-      - finger_r_joint10
-      - finger_r_joint11
-      - finger_r_joint12
-  ring_controller:
-    type: FollowJointTrajectory
-    joints:
-      - finger_r_joint13
-      - finger_r_joint14
-      - finger_r_joint15
-      - finger_r_joint16
-  pinky_controller:
-    type: FollowJointTrajectory
-    joints:
-      - finger_r_joint17
-      - finger_r_joint18
-      - finger_r_joint19
-      - finger_r_joint20
